move xfp stuff from hd.py to descriptor.py
